+++ src/Photo.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Crew\Unsplash;
+namespace Unsplash;
 
 /**
  * Class Photo
- * @package Crew\Unsplash
+ * @package Unsplash
  * @property int $id
  * @property array $user
  */
@@ -45,88 +45,6 @@ class Photo extends Endpoint
         return new ArrayObject($photosArray, $photos->getHeaders());
     }
 
-
-    /**
-     * Retrieve a single page from the list of the curated photos (front-page’s photos).
-     * Returns an ArrayObject that contains Photo objects.
-     *
-     * @param  integer $page Page from which the photos need to be retrieve
-     * @param  integer $per_page Number of element in a page
-     * @param string $order_by Order in which to retrieve photos
-     * @return ArrayObject of Photos
-     */
-    public static function curated($page = 1, $per_page = 10, $order_by = 'latest')
-    {
-        $photos = self::get("/photos/curated", [
-            'query' => ['page' => $page, 'per_page' => $per_page, 'order_by' => $order_by]
-        ]);
-
-        $photosArray = self::getArray($photos->getBody(), get_called_class());
-
-        return new ArrayObject($photosArray, $photos->getHeaders());
-    }
-
-    /**
-     * Retrieve all the photos on a specific page depending on search results
-     * Returns ArrayObject that contain Photo object.
-     *
-     * @param string $search Retrieve photos matching the search term.
-     * @param integer $category Retrieve photos matching the category ID
-     * @param integer $page Page from which the photos need to be retrieved
-     * @param integer $per_page Number of elements on a page
-     * @param string|null $orientation Orientation to search for
-     * @deprecated
-     * @see Search::photos()
-     * @return ArrayObject of Photos
-     */
-    public static function search($search, $category = null, $page = 1, $per_page = 10, $orientation = null)
-    {
-        $photos = self::get(
-            "/photos/search",
-            ['query' => [
-                'query' => $search,
-                'category' => $category,
-                'orientation' => $orientation,
-                'page' => $page,
-                'per_page' => $per_page
-                ]
-            ]
-        );
-
-        $photosArray = self::getArray($photos->getBody(), get_called_class());
-
-        return new ArrayObject($photosArray, $photos->getHeaders());
-    }
-
-    /**
-     * Create a new photo. The user needs to connect their account and authorize the write_photo permission scope.
-     *
-     * @param  string $filePath Path of the file to upload
-     * @throws Exception - if filePath does not exist
-     * @return Photo
-     */
-    public static function create($filePath)
-    {
-        if (!file_exists($filePath)) {
-            throw new Exception(["{$filePath} has not been found"]);
-        }
-
-        $file = fopen($filePath, 'r');
-
-        $photo = json_decode(
-            self::post(
-                "photos",
-                [
-                    'multipart' => [['name' => 'photo', 'contents' => $file]],
-                    'headers' => ['Content-Length' => filesize($filePath)]
-                ]
-            )->getBody(),
-            true
-        );
-
-        return new self($photo);
-    }
-
     /**
      * Retrieve the user that uploaded the photo
      *
@@ -149,10 +67,6 @@ class Photo extends Endpoint
      */
     public static function random($filters = [])
     {
-        if (isset($filters['category']) && is_array($filters['category'])) {
-            $filters['category'] = implode(',', $filters['category']);
-        }
-
         $filters['featured'] = (isset($filters['featured']) && $filters['featured']) ? 'true' : null;
 
         $photo = json_decode(self::get("photos/random", ['query' => $filters])->getBody(), true);