+++ tests/ConnectionTest.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Crew\Unsplash\Tests;
+namespace Unsplash\Tests;
 
-use \Crew\Unsplash as Unsplash;
+use \Unsplash as Unsplash;
 use \League\OAuth2\Client\Token\AccessToken;
 
 /**
  * Class ConnectionTest
- * @package Crew\Unsplash\Tests
+ * @package Unsplash\Tests
  */
 class ConnectionTest extends BaseTest
 {
@@ -16,7 +16,7 @@ class ConnectionTest extends BaseTest
      */
     protected $connection;
 
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();