+++ tests/UserTest.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Crew\Unsplash\Tests;
+namespace Unsplash\Tests;
 
-use \Crew\Unsplash as Unsplash;
+use \Unsplash as Unsplash;
 use \VCR\VCR;
 
 class UserTest extends BaseTest
 {
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();
 
@@ -17,50 +17,44 @@ class UserTest extends BaseTest
 
     public function testFindUser()
     {
-        VCR::insertCassette('users.yml');
+        VCR::insertCassette('users.json');
         $user = Unsplash\User::find('dechuck');
         VCR::eject();
 
         $this->assertEquals('dechuck', $user->username);
     }
 
-    /**
-     * @expectedException \Crew\Unsplash\Exception
-     * @expectedExceptionCode 404
-     */
     public function testFindUnknownUser()
     {
-        VCR::insertCassette('users.yml');
+        $this->expectException(\Unsplash\Exception::class);
+        VCR::insertCassette('users.json');
         Unsplash\User::find('badbadnotgooduser');
         VCR::eject();
     }
 
     public function testFindCurrentUser()
     {
-        VCR::insertCassette('users.yml');
+        VCR::insertCassette('users.json');
         $user = Unsplash\User::current();
         VCR::eject();
 
         $this->assertNotEmpty($user->uploads_remaining);
     }
 
-    /**
-     * @expectedException \Crew\Unsplash\Exception
-     * @expectedExceptionCode 401
-     */
     public function testFindCurrentUserOnUnconnectedUser()
     {
+        $this->expectException(\Unsplash\Exception::class);
         $connection = new Unsplash\Connection($this->provider);
         Unsplash\HttpClient::$connection = $connection;
 
-        VCR::insertCassette('users.yml');
+        VCR::insertCassette('users.json');
         Unsplash\User::current();
         VCR::eject();
     }
 
     public function testFindUserPhotos()
     {
-        VCR::insertCassette('users.yml');
+        VCR::insertCassette('users.json');
         $user = Unsplash\User::find('anniespratt');
         $photos = $user->photos();
         VCR::eject();
@@ -70,7 +64,7 @@ class UserTest extends BaseTest
 
     public function testFindUserPhotosByOrder()
     {
-        VCR::insertCassette('users.yml');
+        VCR::insertCassette('users.json');
         $user = Unsplash\User::find('lukechesser');
         $photos = $user->photos(1, 5, 'oldest');
         VCR::eject();
@@ -80,27 +74,27 @@ class UserTest extends BaseTest
 
     public function testFindUserCollections()
     {
-        VCR::insertCassette('users.yml');
+        VCR::insertCassette('users.json');
         $user = Unsplash\User::find('unsplash');
         $collections = $user->collections();
         VCR::eject();
 
-        $this->assertEquals('Explore Iceland', $collections[0]->title);
+        $this->assertEquals('Unsplash Awards 2019 – Finalists', $collections[0]->title);
     }
 
     public function testFindUserPrivateCollection()
     {
-        VCR::insertCassette('users.yml');
+        VCR::insertCassette('users.json');
         $user = Unsplash\User::current();
         $collections = $user->collections();
         VCR::eject();
 
-        $this->assertEquals('Land', $collections[1]->title);
+        $this->assertEquals('The Final Frontier', $collections[1]->title);
     }
 
     public function testUpdateUser()
     {
-        VCR::insertCassette('users.yml');
+        VCR::insertCassette('users.json');
         $user = Unsplash\User::find('dechuck');
         $user->update(['instagram_username' => 'dechuck123']);
         VCR::eject();
@@ -110,7 +104,7 @@ class UserTest extends BaseTest
 
     public function testFindUserLikedPhoto()
     {
-        VCR::insertCassette('users.yml');
+        VCR::insertCassette('users.json');
         $user = Unsplash\User::find('unsplash', 10, 'oldest');
         $likes = $user->likes();
         $this->assertCount(10, $likes);
@@ -121,7 +115,7 @@ class UserTest extends BaseTest
 
     public function testGetUserPortfolio()
     {
-        VCR::insertCassette('users.yml');
+        VCR::insertCassette('users.json');
         $userPortfolioLink = Unsplash\User::portfolio('hughbertd');
         $this->assertEquals('http://hughbertd.github.io', $userPortfolioLink);
         VCR::eject();
@@ -129,7 +123,7 @@ class UserTest extends BaseTest
 
     public function testUserStatistics()
     {
-        VCR::insertCassette('users.yml');
+        VCR::insertCassette('users.json');
         $user = Unsplash\User::find('hughbertd');
         $statistics = $user->statistics();
         $this->assertArrayHasKey('downloads', $statistics);